chore: enable eth token
