+++ apps/web/src/quoter/utils/crosschain-utils/types.ts
@@ -37,29 +37,3 @@ export interface DestinationSwapQuoteSchema {
   originSwapOrder: InterfaceOrder
   destinationSwapOrder: InterfaceOrder
 }
-
-// OriginChainId -> TokenAddress[]
-export const WHITELIST_TOKEN_MAP = {
-  [ChainId.BSC.toString()]: [
-    // ETH (Binance-Peg Ethereum Token),
-    '0x2170Ed0880ac9A755fd29B2688956BD959F933F8',
-    // USDT
-    '0x55d398326f99059fF775485246999027B3197955',
-    // USDC
-    '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d',
-  ],
-  [ChainId.ARBITRUM_ONE.toString()]: [
-    // USDC
-    '0xaf88d065e77c8cC2239327C5EDb3A432268e5831',
-    // WETH
-    '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
-  ],
-  [ChainId.BASE.toString()]: [
-    // WETH
-    '0x4200000000000000000000000000000000000006',
-    // USDT
-    '0xfde4C96c8593536E31F229EA8f37b2ADa2699bb2',
-    // USDC
-    '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913',
-  ],
-}
